Extract S3RequestSender from S3Wrapper
