+++ extensions/aws/processors/ListS3.h
@@ -58,8 +58,8 @@ class ListS3 : public S3Processor {
     : S3Processor(name, uuid, logging::LoggerFactory<ListS3>::getLogger()) {
   }
 
-  explicit ListS3(std::string name, minifi::utils::Identifier uuid, std::unique_ptr<aws::s3::S3WrapperBase> s3_wrapper)
-    : S3Processor(std::move(name), uuid, logging::LoggerFactory<ListS3>::getLogger(), std::move(s3_wrapper)) {
+  explicit ListS3(std::string name, minifi::utils::Identifier uuid, std::unique_ptr<aws::s3::S3RequestSender> s3_request_sender)
+    : S3Processor(std::move(name), uuid, logging::LoggerFactory<ListS3>::getLogger(), std::move(s3_request_sender)) {
   }
 
   ~ListS3() override = default;
